HorizonServerless/hello_world/requirements.txt
requests
google-cloud-language